update map service test
